+++ Source/Scene/CameraFlightPath.js
@@ -44,7 +44,7 @@ define([
      * <br /><br />
      * Mouse interaction is disabled during flights.
      *
-     * @exports CameraFlightPath
+     * @private
      */
     var CameraFlightPath = {
     };
@@ -127,7 +127,7 @@ define([
         } else {
             points = [ start ];
 
-            angle = Math.acos(Cartesian3.dot(Cartesian3.normalize(afterStart), Cartesian3.normalize(aboveEnd)));
+            angle = CesiumMath.acosClamped(Cartesian3.dot(Cartesian3.normalize(afterStart), Cartesian3.normalize(aboveEnd)));
             axis = Cartesian3.cross(aboveEnd, afterStart);
             if (Cartesian3.equalsEpsilon(axis, Cartesian3.ZERO, CesiumMath.EPSILON6)) {
                 axis = Cartesian3.UNIT_Z;
@@ -388,27 +388,6 @@ define([
     var scratchCartographic = new Cartographic();
     var scratchDestination = new Cartesian3();
 
-    /**
-     * Creates an animation to fly the camera from it's current position to a position given by a Cartesian. All arguments should
-     * be given in world coordinates.
-     *
-     * @param {Scene} scene The scene instance to use.
-     * @param {Object} options Object with the following properties:
-     * @param {Cartesian3} options.destination The final position of the camera.
-     * @param {Cartesian3} [options.direction] The final direction of the camera. By default, the direction will point towards the center of the frame in 3D and in the negative z direction in Columbus view or 2D.
-     * @param {Cartesian3} [options.up] The final up direction. By default, the up direction will point towards local north in 3D and in the positive y direction in Columbus view or 2D.
-     * @param {Number} [options.duration=3000] The duration of the animation in milliseconds.
-     * @param {Function} [options.onComplete] The function to execute when the animation has completed.
-     * @param {Function} [options.onCancel] The function to execute if the animation is cancelled.
-     * @param {Matrix4} [options.endReferenceFrame] The reference frame the camera will be in when the flight is completed.
-     * @param {Boolean} [options.convert=true] When <code>true</code>, the destination is converted to the correct coordinate system for each scene mode. When <code>false</code>, the destination is expected
-     *                  to be in the correct coordinate system.
-     * @returns {Object} An Object that can be added to an {@link AnimationCollection} for animation.
-     *
-     * @exception {DeveloperError} If either direction or up is given, then both are required.
-     *
-     * @see Scene#animations
-     */
     CameraFlightPath.createAnimation = function(scene, options) {
         options = defaultValue(options, defaultValue.EMPTY_OBJECT);
         var destination = options.destination;
@@ -419,15 +398,6 @@ define([
         if (!defined(scene)) {
             throw new DeveloperError('scene is required.');
         }
-        //>>includeEnd('debug');
-
-        if (scene.mode === SceneMode.MORPHING) {
-            return {
-                duration : 0
-            };
-        }
-
-        //>>includeStart('debug', pragmas.debug);
         if (!defined(destination)) {
             throw new DeveloperError('destination is required.');
         }
@@ -435,6 +405,12 @@ define([
             throw new DeveloperError('If either direction or up is given, then both are required.');
         }
         //>>includeEnd('debug');
+
+        if (scene.mode === SceneMode.MORPHING) {
+            return {
+                duration : 0
+            };
+        }
 
         var convert = defaultValue(options.convert, true);
 
@@ -463,9 +439,9 @@ define([
         var onCancel = wrapCallback(options.onCancel);
 
         var camera = scene.camera;
-        var referenceFrame = options.endReferenceFrame;
-        if (defined(referenceFrame)) {
-            camera.setTransform(referenceFrame);
+        var transform = options.endTransform;
+        if (defined(transform)) {
+            camera.setTransform(transform);
         }
 
         var frustum = camera.frustum;
@@ -559,20 +535,6 @@ define([
         };
     };
 
-    /**
-     * Creates an animation to fly the camera from it's current position to a position in which the entire rectangle will be visible. All arguments should
-     * be given in world coordinates.
-     *
-     * @param {Scene} scene The scene instance to use.
-     * @param {Rectangle} options.destination The final position of the camera.
-     * @param {Number} [options.duration=3000] The duration of the animation in milliseconds.
-     * @param {Function} [onComplete] The function to execute when the animation has completed.
-     * @param {Function} [onCancel] The function to execute if the animation is cancelled.
-     * @param {Matrix4} [endReferenceFrame] The reference frame the camera will be in when the flight is completed.
-     * @returns {Object} An Object that can be added to an {@link AnimationCollection} for animation.
-     *
-     * @see Scene#animations
-     */
     CameraFlightPath.createAnimationRectangle = function(scene, options) {
         options = defaultValue(options, defaultValue.EMPTY_OBJECT);
         var rectangle = options.destination;
@@ -581,7 +543,6 @@ define([
         if (!defined(scene)) {
             throw new DeveloperError('scene is required.');
         }
-
         if (!defined(rectangle)) {
             throw new DeveloperError('options.destination is required.');
         }